book_scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pdfkit==1.0.0
python-docx==0.8.11
fpdf
